requirements.txt
boto3==1.17.54
botocore==1.20.54
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.2.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.0
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.0
six==1.15.0
tweepy==3.10.0
urllib3==1.26.4
